requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
gunicorn==21.2.0

